requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.8.0

pandas==2.3.3
scikit-learn==1.7.2
pyyaml==6.0.3
matplotlib==3.10.7
jupyter==1.1.1